+++ RevitAddin.Dockable.Example/Revit/Commands/CommandHide.cs
@@ -29,6 +29,7 @@ namespace RevitAddin.Dockable.Example.Revit.Commands
             UIApplication uiapp = commandData.Application;
 
             App.DockablePaneService.Get(DockablePage.Guid2)?.Hide();
+            App.DockablePaneCreatorService.Get(DockablePage.Guid3)?.Hide();
 
             return Result.Succeeded;
         }