+++ RevitAddin.Dockable.Example/Revit/Commands/CommandShow.cs
@@ -27,6 +27,7 @@ namespace RevitAddin.Dockable.Example.Revit.Commands
             UIApplication uiapp = commandData.Application;
 
             App.DockablePaneService.Get(DockablePage.Guid2)?.Show();
+            App.DockablePaneCreatorService.Get(DockablePage.Guid3)?.Show();
 
             return Result.Succeeded;
         }